Correction and new functionality developer

+++ src/main/java/org/productservice/dto/ProductRequest.java
@@ -1,6 +1,5 @@
 package org.productservice.dto;
 
-import java.math.BigDecimal;
 
-public record ProductRequest(String id,String skuCode, String name,String description,BigDecimal price){
+public record ProductRequest(String id,String skuCode, String name,String description,Double price){
         }
